feat: add check card  for clear cache
