+++ code/handlers/common.go
@@ -1,158 +1,11 @@
 package handlers
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
-	larkcard "github.com/larksuite/oapi-sdk-go/v3/card"
 	"regexp"
-	"start-feishubot/initialization"
 	"strings"
-
-	"github.com/google/uuid"
-	larkim "github.com/larksuite/oapi-sdk-go/v3/service/im/v1"
 )
-
-func replyMarkdown(ctx context.Context, msg string,
-	msgId *string) error {
-	fmt.Println("sendMsg", msg, msgId)
-	msg, i := processMessage(msg)
-	if i != nil {
-		return i
-	}
-	client := initialization.GetLarkClient()
-	config := larkcard.NewMessageCardConfig().
-		WideScreenMode(true).
-		EnableForward(true).
-		UpdateMulti(false).
-		Build()
-
-	// header
-	header := larkcard.NewMessageCardHeader().
-		Template(larkcard.TemplateRed).
-		Title(larkcard.NewMessageCardPlainText().
-			Content("").
-			Build()).
-		Build()
-
-	// Elements
-	divElement := larkcard.NewMessageCardDiv().
-		Fields([]*larkcard.MessageCardField{larkcard.NewMessageCardField().
-			Text(larkcard.NewMessageCardLarkMd().
-				Content(msg).
-				Build()).
-			IsShort(true).
-			Build()}).
-		Build()
-
-	// 谁处理了问题
-	content := "✅ " + "name" + "已处理了此告警"
-	_ = larkcard.NewMessageCardDiv().
-		Fields([]*larkcard.MessageCardField{larkcard.NewMessageCardField().
-			Text(larkcard.NewMessageCardLarkMd().
-				Content(content).
-				Build()).
-			IsShort(true).
-			Build()}).
-		Build()
-
-	// 卡片消息体
-	cardContent, err := larkcard.NewMessageCard().
-		Config(config).
-		Header(header).
-		Elements([]larkcard.MessageCardElement{divElement}).
-		String()
-
-	resp, err := client.Im.Message.Reply(ctx, larkim.NewReplyMessageReqBuilder().
-		MessageId(*msgId).
-		Body(larkim.NewReplyMessageReqBodyBuilder().
-			MsgType(larkim.MsgTypeInteractive).
-			Uuid(uuid.New().String()).
-			Content(cardContent).
-			Build()).
-		Build())
-
-	// 处理错误
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-
-	// 服务端错误处理
-	if !resp.Success() {
-		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
-		return err
-	}
-	return nil
-}
-func replyMsg(ctx context.Context, msg string, msgId *string) error {
-	fmt.Println("sendMsg", msg, msgId)
-	msg, i := processMessage(msg)
-	if i != nil {
-		return i
-	}
-	client := initialization.GetLarkClient()
-	content := larkim.NewTextMsgBuilder().
-		Text(msg).
-		Build()
-
-	resp, err := client.Im.Message.Reply(ctx, larkim.NewReplyMessageReqBuilder().
-		MessageId(*msgId).
-		Body(larkim.NewReplyMessageReqBodyBuilder().
-			MsgType(larkim.MsgTypeText).
-			Uuid(uuid.New().String()).
-			Content(content).
-			Build()).
-		Build())
-
-	// 处理错误
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-
-	// 服务端错误处理
-	if !resp.Success() {
-		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
-		return err
-	}
-	return nil
-}
-func sendMsg(ctx context.Context, msg string, chatId *string) error {
-	//fmt.Println("sendMsg", msg, chatId)
-	msg, i := processMessage(msg)
-	if i != nil {
-		return i
-	}
-	client := initialization.GetLarkClient()
-	content := larkim.NewTextMsgBuilder().
-		Text(msg).
-		Build()
-
-	//fmt.Println("content", content)
-
-	resp, err := client.Im.Message.Create(ctx, larkim.NewCreateMessageReqBuilder().
-		ReceiveIdType(larkim.ReceiveIdTypeChatId).
-		Body(larkim.NewCreateMessageReqBodyBuilder().
-			MsgType(larkim.MsgTypeText).
-			ReceiveId(*chatId).
-			Content(content).
-			Build()).
-		Build())
-
-	// 处理错误
-	if err != nil {
-		fmt.Println(err)
-		return err
-	}
-
-	// 服务端错误处理
-	if !resp.Success() {
-		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
-		return err
-	}
-	return nil
-}
 
 //func sendCard
 func msgFilter(msg string) string {
