+++ code/handlers/msg.go
@@ -1,0 +1,166 @@
+package handlers
+
+import (
+	"context"
+	"fmt"
+	"github.com/google/uuid"
+	"github.com/larksuite/oapi-sdk-go/v3/card"
+	"github.com/larksuite/oapi-sdk-go/v3/service/im/v1"
+	"start-feishubot/initialization"
+)
+
+func sendCard(ctx context.Context, msg string,
+	chatId *string) error {
+	msg, i := processMessage(msg)
+	if i != nil {
+		return i
+	}
+	client := initialization.GetLarkClient()
+	config := larkcard.NewMessageCardConfig().
+		WideScreenMode(false).
+		EnableForward(true).
+		UpdateMulti(false).
+		Build()
+	// header
+	header := larkcard.NewMessageCardHeader().
+		Template(larkcard.TemplateBlue).
+		Title(larkcard.NewMessageCardPlainText().
+			Content("🤖️机器人提醒").
+			Build()).
+		Build()
+
+	// Elements
+	divElement := larkcard.NewMessageCardDiv().
+		Fields([]*larkcard.MessageCardField{larkcard.NewMessageCardField().
+			Text(larkcard.NewMessageCardLarkMd().
+				Content(msg).
+				Build()).
+			IsShort(true).
+			Build()}).
+		Build()
+
+	divElement3 := larkcard.NewMessageCardNote().
+		Elements([]larkcard.MessageCardNoteElement{larkcard.NewMessageCardPlainText().
+			Content("请注意，这将开始一个全新的对话，您将无法利用之前的对话历史信息").
+			Build()}).
+		Build()
+
+	divElement4 := larkcard.NewMessageCardAction().
+		Actions([]larkcard.MessageCardActionElement{
+			larkcard.NewMessageCardEmbedButton().
+				Type(larkcard.MessageCardButtonTypeDanger).
+				Value(map[string]interface{}{"key1": "value1"}).
+				Text(larkcard.NewMessageCardPlainText().
+					Content("确认清除").
+					Build()),
+			larkcard.NewMessageCardEmbedButton().
+				Type(larkcard.MessageCardButtonTypePrimary).
+				Value(map[string]interface{}{"key1": "value1"}).
+				Text(larkcard.NewMessageCardPlainText().
+					Content("我再想想").
+					Build()),
+		}).Layout(larkcard.MessageCardActionLayoutBisected.Ptr()).
+		Build()
+
+	// 卡片消息体
+	cardContent, err := larkcard.NewMessageCard().
+		Config(config).
+		Header(header).
+		Elements([]larkcard.MessageCardElement{divElement, divElement4,
+			divElement3}).
+		String()
+
+	resp, err := client.Im.Message.Create(ctx, larkim.NewCreateMessageReqBuilder().
+		ReceiveIdType(larkim.ReceiveIdTypeChatId).
+		Body(larkim.NewCreateMessageReqBodyBuilder().
+			MsgType(larkim.MsgTypeInteractive).
+			ReceiveId(*chatId).
+			Content(cardContent).
+			Build()).
+		Build())
+
+	// 处理错误
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+
+	// 服务端错误处理
+	if !resp.Success() {
+		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
+		return err
+	}
+	return nil
+}
+func replyMsg(ctx context.Context, msg string, msgId *string) error {
+	fmt.Println("sendMsg", msg, msgId)
+	msg, i := processMessage(msg)
+	if i != nil {
+		return i
+	}
+	client := initialization.GetLarkClient()
+	content := larkim.NewTextMsgBuilder().
+		Text(msg).
+		Build()
+
+	resp, err := client.Im.Message.Reply(ctx, larkim.NewReplyMessageReqBuilder().
+		MessageId(*msgId).
+		Body(larkim.NewReplyMessageReqBodyBuilder().
+			MsgType(larkim.MsgTypeText).
+			Uuid(uuid.New().String()).
+			Content(content).
+			Build()).
+		Build())
+
+	// 处理错误
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+
+	// 服务端错误处理
+	if !resp.Success() {
+		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
+		return err
+	}
+	return nil
+}
+func sendMsg(ctx context.Context, msg string, chatId *string) error {
+	//fmt.Println("sendMsg", msg, chatId)
+	msg, i := processMessage(msg)
+	if i != nil {
+		return i
+	}
+	client := initialization.GetLarkClient()
+	content := larkim.NewTextMsgBuilder().
+		Text(msg).
+		Build()
+
+	//fmt.Println("content", content)
+
+	resp, err := client.Im.Message.Create(ctx, larkim.NewCreateMessageReqBuilder().
+		ReceiveIdType(larkim.ReceiveIdTypeChatId).
+		Body(larkim.NewCreateMessageReqBodyBuilder().
+			MsgType(larkim.MsgTypeText).
+			ReceiveId(*chatId).
+			Content(content).
+			Build()).
+		Build())
+
+	// 处理错误
+	if err != nil {
+		fmt.Println(err)
+		return err
+	}
+
+	// 服务端错误处理
+	if !resp.Success() {
+		fmt.Println(resp.Code, resp.Msg, resp.RequestId())
+		return err
+	}
+	return nil
+}
+
+func sendClearCacheCheckCard(ctx context.Context, chatId *string) {
+	sendCard(ctx, "您确定要清除对话上下文吗？", chatId)
+}
